move button state handlers to separate files, for code optimization
