+++ client/src/helpers/wishlistButtonStateHandler.js
@@ -1,0 +1,22 @@
+/* eslint-disable no-underscore-dangle */
+const wishlistButtonStateHandler = (
+  globalUserState,
+  globalUserDataState,
+  localButtonState,
+  localButtonStateSetter,
+  productId,
+) => {
+  if (globalUserState) {
+    if (globalUserDataState.length > 0) {
+      localButtonStateSetter(
+        globalUserDataState.some((p) => p._id === productId),
+      );
+    } else if (localButtonState) {
+      localButtonStateSetter(!localButtonState);
+    }
+  } else {
+    localButtonStateSetter(false);
+  }
+};
+
+export default wishlistButtonStateHandler;
